Added Yes/no to deletion options

+++ CTS_Application/frmSettings.cs
@@ -12,7 +12,8 @@ namespace CTS_Application
 {
     public partial class frmSettings : Form
     {
-        DbConnect con = new DbConnect();
+        //The DbConnect class cannot be declared globaly. If one want to delete several tables it will only work if its declared in each event.
+        
         
         public frmSettings()
         {
@@ -21,19 +22,44 @@ namespace CTS_Application
 
         private void btnDelRec_Click(object sender, EventArgs e)
         {
-            con.DeleteRecordsInTable("historian");
+            DialogResult dialogResult = MessageBox.Show("This will delete all temperature recordings", "WARNING!", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                DbConnect con = new DbConnect();
+                con.DeleteRecordsInTable("historian");
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+                //do something else
+            }
         }
 
         private void btnDelAla_Click(object sender, EventArgs e)
         {
-            con.DeleteRecordsInTable("alarm_historian");
+            DialogResult dialogResult = MessageBox.Show("This will delete all alarm recordings", "WARNING!", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                DbConnect con = new DbConnect();
+                con.DeleteRecordsInTable("alarm_historian");
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+                //do something else
+            }
         }
 
         private void btnDelSub_Click(object sender, EventArgs e)
         {
-            con.DeleteRecordsInTable("users");
+            DialogResult dialogResult = MessageBox.Show("This will delete all subscribers", "WARNING!", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                DbConnect con = new DbConnect();
+                con.DeleteRecordsInTable("users");
+            }
+            else if (dialogResult == DialogResult.No)
+            {
+                //do something else
+            }
         }
-
-    
     }
 }
